Move stage-to-prime fileset test to test_prime.py
